% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositional_biwm_krig.R
\name{compositional_biwm_krig}
\alias{compositional_biwm_krig}
\title{Compositional Krigging for the Bivariate (Wittle-)Matern Model.}
\usage{
compositional_biwm_krig(
  biwm_fit,
  krig_locations,
  fit_locations,
  obs_matrix,
  nus,
  nug_vec
)
}
\arguments{
\item{biwm_fit}{An object of biwm class. Must be the result of a `fit_biwm` command.}

\item{krig_locations}{A n x 2 numeric matrix of coordinates for the unobserved data.}

\item{fit_locations}{A n x 2 numeric matrix of coordinates for the estimation data.}

\item{obs_matrix}{A n x 2 numeric matrix of data. We assume that the user has already alr-transformed the data.}

\item{nus}{A 2 x 1 numeric vector with the smoothness coefficients for the Mat\'{e}rn covariance function, c(nu1, nu2).}
}
\value{
\code{compositional_biwm_krig} returns a matrix with alr-transformed data for given krigging locations.
}
\description{
Compositional Krigging for the Bivariate (Wittle-)Matern Model.
}
\examples{
set.seed(3)
n <- 40
coords <- matrix(runif(2*n), ncol = 2)
temp <- rnorm(2*n)
nus_vec <- c(0.5, 0.5)
nug_vec <- c(0,0)
S <- sigma_assembler_biwm(sigmas = c(1, 1), a = 2, rho = 0.5,
nus = nus_vec, coords_matrix = coords, combined = TRUE,nug_vec = nug_vec)
log_cd <- matrix(temp\%*\%chol(S) + rep(c(1,2), each = n), ncol = 2)

informed_test <- fit_biwm(log_cd, coords, c(1, 1, 2, .5), nus_vec,nug_vec = nug_vec)
biwm_fit <- fit_biwm(log_cd, coords, c(.5, .5, 4, .6), nus_vec,nug_vec = nug_vec)

x_grid <- seq(0,1,length.out = 10)
krig_locations <- as.matrix(expand.grid(x_grid,x_grid))

compositional_biwm_krig(biwm_fit,krig_locations,coords,log_cd,nus_vec,nug_vec)

}
