% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_cov_mat.R
\name{conditional_cov_mat}
\alias{conditional_cov_mat}
\title{Compute the Conditional Multivariate Normal Covariance Matrix}
\usage{
conditional_cov_mat(loc_new, loc_obs, sigmas, a, rho, nus, nug_vec)
}
\description{
Compute the Conditional Multivariate Normal Covariance Matrix
}
\examples{

set.seed(1234)
loc_obs <- matrix(runif(2*10),ncol =2)
loc_new <- matrix(runif(2*3),ncol =2)

sigmas <- c(1,2)
a <- 0.3
rho <- 0.8
nus <- c(1,1)
nug_vec <- c(0,0)


std_err_mat <- conditional_cov_mat(loc_new,loc_obs,sigmas, a, rho, nus, nug_vec)


}
