% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_fun_case_01.R
\name{experiment_fun_case_01}
\alias{experiment_fun_case_01}
\title{Experimentation with fit_biwm Case 01: Starting values are the true theta parameters. In this case, we're constrained to nu_1 = 0.5 and nu_2 = 0.5.}
\usage{
experiment_fun_case_01(grid, coords)
}
\arguments{
\item{grid}{A n x 4 numeric matrix with theta points. The columns represent the theta parameters used in the fit_biwm function.}

\item{coords_matrix}{A n x 2 numeric matrix of coordinates for the data.}
}
\value{
\code{experiment_fun_case_01} returns a list whose elements are lists that contains the output of fit_biwm some more.
  \item{mu}{Expected value estimate}
  \item{theta}{Estimated Value by Maximum Likelihood}
  \item{likelihood}{Likelihood as proposed by the model}
  \item{start}{Starting values used in estimation.}
  \item{diff}{Difference between estimated and true values.}
}
\description{
Experimentation with fit_biwm Case 01: Starting values are the true theta parameters. In this case, we're constrained to nu_1 = 0.5 and nu_2 = 0.5.
}
\examples{

# ------ NOT RUN ------

set.seed(3)
n <- 40
coords <- matrix(runif(2*n), ncol = 2)
temp <- rnorm(2*n)

sigma_1 <- seq(0.5,5,length.out = 20)
sigma_2 <- seq(0.5,5,length.out = 20)
as <- seq(0.5,2,length.out = 20)
rho = 0.5
grid_01 <- expand.grid(sigma1 = sigma_1,sigma2 = sigma_2,a = as,rho = rho)
result <- experiment_fun_case_01(grid_01,coords)
}
