% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_biwm.R
\name{fit_biwm}
\alias{fit_biwm}
\title{Log-likelihood of the Bivariate (Wittle-)Matern Model.}
\usage{
fit_biwm(obs_vec, coords_matrix, theta0, nus, verbosity = 0, ...)
}
\arguments{
\item{obs_vec}{A n x 2 numeric matrix of data. We assume that the user has already log-transformed the data.}

\item{coords_matrix}{A n x 2 numeric matrix of coordinates for the data.}

\item{theta0}{Initial values for the optimizer. A 4 x 1 numeric vector, the order of parameters is: theta = c(sigma_1^2, sigma_2^2, a, rho).}

\item{nus}{A 2 x 1 numeric vector with the smoothness coefficients for the Mat\'{e}rn covariance function, c(nu1, nu2).}
}
\value{
\code{fit_biwm} returns an object of \code{\link{class}} "biwm" containing at least the following components
  \item{mu}{Expected value estimate}
  \item{theta}{TODO}
  \item{se.mu}{TODO}
  \item{se.theta}{TODO}
  \item{likelihood}{TODO}
}
\description{
Log-likelihood of the Bivariate (Wittle-)Matern Model.
}
\examples{
set.seed(3)
n <- 40
coords <- matrix(runif(2*n), ncol = 2)
temp <- rnorm(2*n)
S <- sigma_assembler_biwm(sigmas = c(1, 1), a = 2, rho = 0.5, nus = c(0.5, 0.5), coords_matrix = coords)
log_cd <- matrix(temp\%*\%chol(S) + rep(c(1,2), each = n), ncol = 2)

informed_test <- fit_biwm(log_cd, coords, c(1, 1, 2, .5), c(0.5, 0.5))
generic_test <- fit_biwm(log_cd, coords, c(.5, .5, 4, .6), c(0.5, 0.5))

}
