% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_approx_alrInv_cond_covvar.R
\name{get_approx_alrInv_cond_covvar}
\alias{get_approx_alrInv_cond_covvar}
\title{Obtain Prediction Standard Error Estimate}
\usage{
get_approx_alrInv_cond_covvar(
  krig_values,
  loc_new,
  loc_obs,
  biwm_fit,
  nus,
  nug_vec
)
}
\arguments{
\item{krig_values}{A N x 2 numeric matrix of krig values not in the Simplex.}

\item{loc_new}{A N x 2 numeric matrix of coordinates for the unobserved data.}

\item{loc_obs}{A n x 2 numeric matrix of coordinates for the estimation data.}

\item{biwm_fit}{An object of biwm class. Must be the result of a `fit_biwm` command.}

\item{nus}{A 2 x 1 numeric vector with the smoothness coefficients for the Mat\'{e}rn covariance function, c(nu1, nu2).

#' @param nug_vec A 2 x 1 numeric vector with nugget effects Mat\'{e}rn covariance function, c(nu1, nu2).}
}
\description{
Obtain Prediction Standard Error Estimate
}
\examples{
set.seed(3)
n <- 40
coords <- matrix(runif(2*n), ncol = 2)
temp <- rnorm(2*n)
nug_vec <- c(0,0)
nus <- c(0.5,0.5)
S <- sigma_assembler_biwm(sigmas = c(1, 1), a = 2, rho = 0.5,
nus = nus, coords_matrix = coords, nug_vec = nug_vec, combined = TRUE)
log_cd <- matrix(temp\%*\%chol(S) + rep(c(1,2), each = n), ncol = 2)

generic_test <- fit_biwm(log_cd, coords, c(.5, .5, 4, .6), nus = nus,nug_vec = nug_vec)

loc_new <- matrix(runif(2*20),ncol =2)

pred_std_err_mat <- get_pred_std_err(loc_new = loc_new, loc_obs = coords,biwm_fit = generic_test, nus = nus, nug_vec = nug_vec)


krig_vals <-
  compositional_biwm_krig(
    biwm_fit = generic_test,
    krig_locations = loc_new,
    fit_locations = coords,
    obs_matrix = log_cd,
    nus = nus,
    nug_vec = nug_vec,
    return_comp = FALSE
  )

get_approx_alrInv_cond_covvar(krig_values = krig_vals,loc_new = loc_new, loc_obs = coords,biwm_fit = generic_test, nus = nus, nug_vec = nug_vec)




}
