% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLike_biwm.R
\name{LLike_biwm}
\alias{LLike_biwm}
\title{Log-likelihood of the Bivariate (Wittle-)Matern Model.}
\usage{
LLike_biwm(theta, nus, mu, coords_matrix, obs_vec, parsimonious = TRUE, ...)
}
\value{
log-likelihood
}
\description{
Log-likelihood of the Bivariate (Wittle-)Matern Model.
}
