% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loglike_contour_field.R
\name{plot_loglike_contour_ggplot}
\alias{plot_loglike_contour_ggplot}
\title{Likelihood Contour Plot for the Bivariate (Wittle-)Matern Model.}
\usage{
plot_loglike_contour_ggplot(
  var_1,
  var_2,
  grid_df,
  obs,
  coords,
  true_param = NULL,
  nbins = 30,
  gamma_plot = 1
)
}
\description{
Likelihood Contour Plot for the Bivariate (Wittle-)Matern Model.
}
\examples{

n <- 70
coords <- matrix(runif(2*n), ncol = 2)
temp <- rnorm(2*n)
nus_vec <- c(0.5,0.5)

true_theta = c(1,1,2,0.5)
sigma2_1_grid <- seq(0.5,2,length.out = 15)
sigma2_2_grid <- seq(0.5,2,length.out = 15)
a_grid <- seq(0.5,3,length.out = 15)
rho_grid <- seq(0.25,0.75,length.out = 15)

S <- sigma_assembler_biwm(sigmas = true_theta[1:2],
                          a = true_theta[3],
                          rho = true_theta[4],
                          nus = c(0.5, 0.5),
                          coords_matrix = coords,
                          combined = TRUE)

log_cd <- matrix(temp\%*\%chol(S) + rep(c(1,2), each = n), ncol = 2)

grid_01 <- construct_grid(
  sigma2_1_grid,
  1,
  a_grid,
  0.2)

plot_loglike_contour_ggplot("sigma2_1","a",grid_df = grid_01,obs = log_cd,coords = coords,nbins = 30,gamma_plot =1e3,true_param = true_theta[c(1,3)])


grid_02 <- construct_grid(
  1,
  1,
  a_grid,
  rho_grid
)

plot_loglike_contour_ggplot("a","rho",grid_df = grid_02,obs = log_cd,coords = coords,nbins = 30,gamma_plot =1e3,true_param = true_theta[c(3,4)])



grid_03 <- construct_grid(
  sigma2_1_grid,
  sigma2_2_grid,
  2,
  0.5)

plot_loglike_contour_ggplot("sigma2_1","sigma2_2",grid_df = grid_03,obs = log_cd,coords = coords,nbins = 30,gamma_plot =1e3,true_param = true_theta[c(1,2)])




grid_04 <- construct_grid(
  sigma2_1_grid,
  1,
  1,
  rho_grid
)

plot_loglike_contour_ggplot("sigma2_1","rho",grid_df = grid_04,obs = log_cd,coords = coords,nbins = 30,gamma_plot =1e3,true_param = true_theta[c(1,4)])
}
