% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results_case_01.R
\name{plot_results}
\alias{plot_results}
\title{Plot the results of experiment_fun_case_01: Estimated Value vs True Value and Difference vs True Value. Here you can choose to either plot them in the same device or to have them plotted separately.}
\usage{
plot_results(results, display_both = TRUE)
}
\arguments{
\item{results}{The result of the experiment_fun_case_01 function. It should be a list of lists, whose second hierarchy level is the output of fit_biwm.}

\item{display_both}{Logical. Should the plots be in the same device or not.}
}
\description{
Plot the results of experiment_fun_case_01: Estimated Value vs True Value and Difference vs True Value. Here you can choose to either plot them in the same device or to have them plotted separately.
}
\examples{
set.seed(3)
n <- 40
coords <- matrix(runif(2*n), ncol = 2)
temp <- rnorm(2*n)

sigma_1 <- seq(0.5,5,length.out = 20)
sigma_2 <- seq(0.5,5,length.out = 20)
as <- seq(0.5,2,length.out = 20)
rho = 0.5
grid_01 <- expand.grid(sigma1 = sigma_1,sigma2 = sigma_2,a = as,rho = rho)
result <- experiment_fun_case_01(grid_01[sample(1:8e3,5),],coords)
plot_results(result)
}
