% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_and_estimate_biwm.R
\name{simulate_and_estimate_biwm}
\alias{simulate_and_estimate_biwm}
\title{Compositional Krigging Simulation for the Bivariate (Wittle-)Matern Model.}
\usage{
simulate_and_estimate_biwm(
  n_points,
  true_params,
  initial_params,
  nus_vec = c(0.5, 0.5),
  seed_number = 123
)
}
\description{
Compositional Krigging Simulation for the Bivariate (Wittle-)Matern Model.
}
\examples{

initial_point <- c(0.2756810, 0.7721986, 2.9531961, 0.9619193)
sim_result <- simulate_and_estimate_biwm(n_points = 80,
true_params = c(1,3,1,0.9),
initial_params = initial_point,
seed_number = sample.int(123123,1),
nus_vec = c(1,1.5)
)
}
