% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_hat_distrib_estim.R
\name{theta_hat_hist}
\alias{theta_hat_hist}
\title{Bivariate (Wittle-)Matern Model Paramaters Histogram Simulation.}
\usage{
theta_hat_hist(
  true_param,
  initial_guess,
  n_points = 80,
  nus_vec = c(0.5, 0.5),
  n_replicates = 30,
  seed_number = sample.int(12312312, 1)
)
}
\arguments{
\item{true_param}{A length 4 vector of true parameters.}

\item{initial_guess}{A length 4 vector of initial points for the optimization.}

\item{n_points}{Number of spatial locations.}

\item{nus_vec}{A 2 x 1 numeric vector with the smoothness coefficients for the Mat\'{e}rn covariance function, c(nu1, nu2).}

\item{n_replicates}{Number of experiment replications.}

\item{seed_number}{seed.}
}
\description{
Fixes spatial coordinates and simulate the Bivariate Normal observations. Then estimate the true parameters density with histogram.
}
\examples{

true_param <- c(1.5,0.7,1.3,-0.4)
initial_guess <-  c(0.5,5,4,0.2)
theta_hat_hist(true_param,initial_guess,n_replicates = 50)
}
