% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLike_biwm.R
\name{LLike_biwm}
\alias{LLike_biwm}
\title{Log-likelihood of the Bivariate (Wittle-)Matern Model.}
\usage{
LLike_biwm(theta, nus, mu, coords_matrix, obs_vec)
}
\arguments{
\item{theta}{A 4 x 1 numeric vector, the order of parameters is: theta = c(sigma_1^2, sigma_2^2, a, rho).}

\item{nus}{A 2 x 1 numeric vector with the smoothness coefficients for the Mat\'{e}rn covariance function, c(nu1, nu2).}

\item{mu}{A 2 x 1 numeric vector of means c(mu1, mu2).}

\item{coords_matrix}{A n x 2 numeric matrix of coordinates for the data.}

\item{obs_vec}{A n x 2 numeric matrix of data. We assume that the user has already log-transformed the data.}
}
\value{
log-likelihood
}
\description{
Log-likelihood of the Bivariate (Wittle-)Matern Model.
}
